RX-2997: support FAIL_FAST flag for casper-tests

+++ build-server/test.coffee
@@ -9,24 +9,22 @@ unlink = Promise.promisify unlink
 {TEST_REPORTS}=process.env
 CODE_SIGNALLED = 98
 CODE_REPORT_MISSING = 2
+FAIL_FAST = process.env['FAIL_FAST'] isnt '0'
 
 existsAsync = (path) ->
     new Promise (resolve) -> exists path, resolve
 
-prepareTest = Promise.coroutine (makeTarget, reportFile, env) ->
-    if env?
-        for k, v in process.env
-            env[k] = v
-    else
-        env = process.env
+prepareTest = Promise.coroutine (makeTarget, reportFile, env = {}) ->
+    for k, v in process.env
+        env[k] = v
 
-    yield mkdirp path.dirname reportFile
     env['MAKE_TARGET'] = makeTarget
     env['REPORT_FILE'] = reportFile
     if TEST_REPORTS?
         env['PREFIX'] = TEST_REPORTS
         reportFile = path.join TEST_REPORTS, reportFile
 
+    yield mkdirp path.dirname reportFile
     e = yield existsAsync reportFile
     yield unlink reportFile if e
 
@@ -44,6 +42,7 @@ exitCallback = ->
 
 cleanup = Promise.coroutine (reportFile, code) ->
     if yield existsAsync reportFile
+        return 0 unless FAIL_FAST
         yield unlink reportFile unless code is 0
     else
         console.error 'Test Report file %s missing', reportFile
@@ -57,16 +56,4 @@ processTest = Promise.coroutine (child, reportFile) ->
     code = yield cb.Promise
     return cleanup reportFile, code
 
-prefix = (reportFile) ->
-    if TEST_REPORTS?
-        _prefix = path.join TEST_REPORTS, 'x'
-        _prefix = _prefix.substring 0, _prefix.length-1
-        if reportFile.substring(0, _prefix.length) is _prefix
-            return _prefix
-
-plainReportFile = (reportFile) ->
-    _prefix = prefix reportFile
-    return reportFile unless _prefix?
-    reportFile.substring _prefix.length
-
-module.exports = {prepareTest, processTest, cleanup, exitCallback, plainReportFile, prefix}
+module.exports = {prepareTest, processTest}
